+++ be/api/userApi/pimpinanDesaAPI.js
@@ -8,6 +8,5 @@ Router.get('/get/:role', getKades);
 Router.post('/create', postKadesWakades);
 Router.put('/update/:role', updateKadesByRole);
 Router.delete('/delete/:id', deleteKadesById);
-Router.put('/submit/:kadesId/:suratAcaraId', SubmitSuratKades);
 
 module.exports = Router;
