+++ be/controllers/adminController/kepalaDesaController.js
@@ -151,62 +151,3 @@ exports.deleteKadesById = async (req, res) => {
         });
     }
 }
-
-exports.SubmitSuratKades = async (req, res) => {
-    const session = await mongoose.startSession();
-    session.startTransaction();
-    try{
-        const {kadesId, suratAcaraId} = req.params;
-        const { statusPersetujuanReq } = req.body;
-        const dataKades = await KadesModel.findById([kadesId]).session(session);
-        if (!dataKades){
-            await session.abortTransaction();
-            return res.status(404).send({
-                message: "Data kades not found"
-            });
-        }
-        const dataSuratAcara = await SuratAcaraModel.findById([suratAcaraId]).session(session);
-        if (!dataSuratAcara){
-            await session.abortTransaction();
-            return res.status(404).send({
-                message: "Data surat acara not found"
-            });
-        }
-        if (dataSuratAcara.statusPersetujuan === "disetujui"){
-            await session.abortTransaction();
-            return res.status(400).send({
-                message: "Surat acara sudah disetujui"
-            });
-        }
-
-        if (statusPersetujuanReq === true && dataSuratAcara.statusAcara === "pengajuan kades dan wakades" && dataSuratAcara.statusPersetujuan === "disetujui perangkat desa"){
-            dataSuratAcara.statusPersetujuan = "disetujui pimpinan desa";
-            dataSuratAcara.statusAcara = "pengajuan selesai";
-            dataKades.suratAcaraApproved.push(dataSuratAcara._id);
-            const indexData = dataKades.suratAcaraPending.indexOf(dataSuratAcara._id);
-            dataKades.suratAcaraPending.splice(indexData, 1);
-            await dataKades.save();
-            await dataSuratAcara.save();
-            await session.commitTransaction();
-            return res.status(200).send({
-                message: "Success submit surat kades",
-                data: dataSuratAcara
-            });
-        }
-
-        res.send({
-            message: "test",
-            data: dataKades,
-            data2: dataSuratAcara
-        });
-
-
-    }catch(error){
-        await session.abortTransaction();
-        res.status(500).send({
-            message: error.message || "Some error occurred while submit surat kades."
-        });
-    }finally{
-        session.endSession();
-    }
-};
