+++ be/controllers/adminController/perangkatDesaController.js
@@ -122,55 +122,3 @@ exports.deletePerangkatDesaById = async (req,res) => {
 }
 
 //merubah statusPersetujuan pada suratAcara menjadi disetujui perangkat desa
-exports.SubmitSuratAcara = async (req,res) => {
-    const { suratAcaraId, perangkatDesaId } = req.params;
-    const { statusPersetujuanReq } = req.body;
-    try{
-        const dataPD = await perangkatDesa.findById(perangkatDesaId);
-        const surat = await SuratAcaraModel.findById(suratAcaraId);
-        
-
-        if (!dataPD || !surat) {
-            return res.status(404).send({
-                message: "perangkat desa or surat acara not found with id " + id
-            });
-        }
-
-        if (surat.statusPersetujuan === 'belum ada persetujuan' || surat.statusPersetujuan === 'ditolak rw') {
-            return res.status(404).send({
-                message: "belum ada persetujuan dari rw " + id
-            });
-        }
-
-        if (surat.statusPersetujuan === 'disetujui rw' && statusPersetujuanReq === true) {
-            surat.statusPersetujuan = 'disetujui perangkat desa';
-            surat.statusAcara = 'pengajuan kades dan wakades';
-
-            dataPD.suratAcaraApproved.push(suratAcaraId);
-            const indexData = dataPD.suratAcaraPending.indexOf(suratAcaraId);
-            dataPD.suratAcaraPending.splice(indexData, 1);
-            await dataPD.save();
-            
-            const DataPimpinanDesa = await PimpinanDesaModel.findOne({rolePemimpinDesa:1});
-            if(!DataPimpinanDesa){
-                return res.status(404).send({message: "kepala desa not found"});
-            } 
-            DataPimpinanDesa.suratAcaraPending.push(suratAcaraId);
-            const indexDataPimpinanDesa = DataPimpinanDesa.suratAcaraComing.indexOf(suratAcaraId);
-            DataPimpinanDesa.suratAcaraComing.splice(indexDataPimpinanDesa, 1);
-            await DataPimpinanDesa.save();
-            surat.pimpinanDesaId = DataPimpinanDesa._id;
-            await surat.save();
-            return res.send({message: "test",result: surat});
-        }
-        else{
-            surat.statusPersetujuan = 'ditolak perangkat desa';
-            await surat.save();
-            return res.send({message: "test",result: surat});
-        }
-    }catch(error){
-        res.status(500).send({
-            message: error.message || "Some error occurred while pelayanan konter by id."
-        });
-    }
-};
