+++ be/controllers/userController/rwController.js
@@ -132,189 +132,4 @@ exports.deleteRwById = async (req,res)=>{
     }
 }
 
-const getRolePd = (jenisSurat) => {
-    const pd1 = [
-        "surat keterangan domisili",
-        "surat keterangan usaha",
-        "surat kenal lahir",
-        "surat pindah",
-        "surat pengantar skck",
-        "surat izin keramaian",
-        "surat izin pepergian",
-    ];
-
-    const pd2 = [
-        "pembuatan kk baru",
-        "pembuatan ktp baru",
-        "pembuatan ktp lama",
-        "pemecahan kk lama",
-        "pbb-p2",
-        "mutasi pbb",
-        "pencatatan kependudukan",
-    ];
-
-    const pd3 = [
-        "surat kematian",
-        "surat kelahiran",
-        "santunan kematian",
-        "pengajuan jkn-kis",
-        "koordinator rtlh",
-        "pendataan masalah sosial",
-        "bantuan sosial",
-    ];
-    if (pd1.includes(jenisSurat)) return 1;
-    if (pd2.includes(jenisSurat)) return 2;
-    if (pd3.includes(jenisSurat)) return 3;
-
-}
-
-const getKasiType = (rolePd) => {
-    const kasiTypes = ["pelayanan", "pemerintahan", "kersa"];
-    return kasiTypes[rolePd - 1] || ""; 
-};
-
-exports.persetujuanSurat = async (req, res) => {
-    const { SuratId, RwId } = req.params;
-    const { statusPersetujuanReq } = req.body;
-
-    try {
-        const PakRw = await RwModel.findById(RwId);
-        const surat = await SuratAcaraModel.findById(SuratId);
-
-        if (!PakRw || !surat) {
-            return res.status(404).send({
-                message: "RW or Surat not found."
-            });
-        }
-
-        if (surat.statusPersetujuan === "belum ada persetujuan" || surat.statusPersetujuan === "ditolak rt") {
-            return res.status(404).send({
-                message: "Belum ada persetujuan dari RT."
-            });
-        }
-
-        if (surat.statusPersetujuan === "disetujui rt" && surat.statusAcara === "pengajuan rw") {
-            if (statusPersetujuanReq === true) {
-                surat.statusPersetujuan = "disetujui rw";
-                const rolePd = getRolePd(surat.jenisSurat);
-
-                if (rolePd) {
-                    surat.statusAcara = `pengajuan perangkat desa kasi ${getKasiType(rolePd)}`;
-                    
-                    const Kasi = await PerangkatDesaModel.findOne({ rolePD: rolePd });
-
-                    if (Kasi) {
-                        Kasi.suratAcaraPending.push(surat._id);
-                        // menghapus surat dari suratAcaraComing
-                        const indexDataKasi = Kasi.suratAcaraComing.indexOf(surat._id);
-                        Kasi.suratAcaraComing.splice(indexDataKasi, 1);
-                        await Kasi.save();
-
-                        surat.rwId = RwId;
-                        await surat.save();
-
-                        PakRw.suratAcaraApproved.push(surat._id);
-                        const indexData = PakRw.suratAcaraPending.indexOf(surat._id);
-                        PakRw.suratAcaraPending.splice(indexData, 1);
-                        await PakRw.save();
-
-                        return res.status(200).send({
-                            message: "Success update persetujuan surat",
-                            info: "Surat sudah disetujui RW dan sudah diajukan ke Perangkat Desa",
-                            data: surat
-                        });
-
-                        
-                    } else {
-                        return res.status(404).send({
-                            message: "Perangkat Desa not found."
-                        });
-                    }
-                }
-
-            } else {
-                surat.statusPersetujuan = "ditolak rw";
-                await surat.save();
-                return res.status(200).send({
-                    message: "Success update persetujuan surat",
-                    info: "Surat ditolak RW",
-                    data: surat
-                });
-            }
-        }
-    } catch (error) {
-        return res.status(500).send({
-            message: error.message || "Some error occurred while updating persetujuan surat."
-        });
-    }
-};
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 module.exports = exports;
